Refactoring of GroupController, task: task-JDP190701-2

+++ src/main/java/com/kodilla/ecommercee/dto/ProductDto.java
@@ -8,9 +8,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class ProductDto {
-    private  int id;
+    private  Long id;
     private String name;
     private String description;
     private double price;
-    private int groupId;
+    private Long groupId;
 }
